pyproject.toml
[project]
name = "job-rank"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.23",
    "langchain-openai>=0.3.14",
    "markdown>=3.8",
    "python-dotenv>=1.1.0",
    "scikit-learn>=1.6.1",
    "sentence-transformers>=4.1.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "mypy>=1.15.0",
    "pre-commit>=4.2.0",
    "ruff>=0.11.6",
]
